Converts search to new Pagefinder
